+++ src/provisionedmachine/HISTORY.rst
@@ -5,6 +5,7 @@ Release History
 1.0.0b7
 ++++++
 * ``az provisionedmachine os-image list`` - Added an ``Architecture`` column to the table output, surfacing the CPU architecture (amd64/arm64) of each OS image.
+* ``az provisionedmachine os-image list`` - Added a ``SolutionType`` column to the table output, surfacing the OS family (HCI/AzureLinux) returned by the service.
 
 1.0.0b6
 ++++++